requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
openai>=1.30.0
anthropic>=0.28.0
google-genai>=1.0.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
pydantic>=2.7.0
rich>=13.7.0
